+++ core/ansible/base.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/TheManticoreProject/gopengraph/node"
-	"github.com/TheManticoreProject/gopengraph/properties"
+	"github.com/Ramoreik/gopengraph/node"
+	"github.com/Ramoreik/gopengraph/properties"
 )
 
 var Instance AnsibleInstance